Orion5.p carries element angles in degrees without gearing applied and sends 'cytron degrees' over the serial.
Now the visualiser/controller has the gearbox scaler for the shoulder built into it, this is (1+(52/28)) and we will have to put this in clear wording somewhere as well into the basic matlab implementation. This should help stop the users overrunning the absolute limits and prevent an:
advantaged offset situation, where it thinks 120deg is actually 0deg cause it rolled over... which could be dangerous to the hardware in terms of pulling cabling.